Moved the safe redirect to common.py
