+++ website/misc/common.py
@@ -1,0 +1,20 @@
+import flask
+import logging
+from urllib.parse import urlparse, urljoin
+from flask import request, url_for
+
+logger = logging.getLogger(__name__)
+
+# http://flask.pocoo.org/snippets/62/
+def is_safe_url(target):
+    ref_url = urlparse(request.host_url)
+    test_url = urlparse(urljoin(request.host_url, target))
+    return test_url.scheme in ('http', 'https') and \
+           ref_url.netloc == test_url.netloc
+
+
+def redirect_back(endpoint, **values):
+    target = request.form['next']
+    if not target or not is_safe_url(target):
+        target = url_for(endpoint, **values)
+    return flask.redirect(target)
